Adjusted ranking to new API. Fixed issues from CR.
