--- src/actions/ranking.actions.js
@@ -1,7 +1,0 @@
-import * as types from './ranking.types';
-
-export const sortBy = (column, isAscendingOrder = true) => ({
-    type: types.SORT,
-    column,
-    isAscendingOrder
-});
