--- src/actions/ranking.types.js
@@ -1,1 +1,0 @@
-export const SORT = 'RANKING::SORT';
